adicionando entidades de venda

+++ src/main/java/com/chrisdev/eng/pdvediel/entity/Item.java
@@ -4,6 +4,8 @@ import jakarta.persistence.*;
 import lombok.Getter;
 import lombok.Setter;
 
+import java.math.BigDecimal;
+
 @Entity
 @Table(name = "itens")
 @Getter
@@ -21,7 +23,7 @@ public class Item {
     private String descricao;
 
     @Column(nullable = false)
-    private Double preco;
+    private BigDecimal preco;
 
     @Enumerated(EnumType.STRING)
     @Column(nullable = false)
